Add Servo experiments
This now allows to send positions to a connected Servo as well as read
its position from the potentiometer feedback wire. Based on this, it
will detach the servo once the position has been reached to a satisfying
degree.

Also, when manually rotating the servo to more than 90 degrees, an
indicator pixel will light up :)

Co-authored-by: Marta83 <martaruizferna@gmail.com>