Add abort functionality
Sometimes we are forgetful and forget to send a message. Or
forget to mark a message as read. Or we make typos and don't
want to send our message like that.

In those cases, we might want to reset the device.

This commit adds two ways of resetting. One is that after 120
seconds of inactivity, our device just returns back to the
idle state and resets its variables and the flag position.

The other is that anyone can send an abort signal to all
devices by pressing the button for 5 seconds. This will reset
everyone's device back to the idle state. This can also be
used to abort while being in the typing mode.